refactor(data-point-service): refactors modules to prepare for new revalidation feature
re #303
