+++ packages/data-point-service/lib/redis-controller.test.js
@@ -1,0 +1,110 @@
+/* eslint-env jest */
+
+const RedisController = require('./redis-controller')
+
+describe('createSWRStaleKey', () => {
+  it('should create stale key', () => {
+    expect(RedisController.createSWRStaleKey('key')).toEqual('key:swr.stale')
+  })
+})
+
+describe('createSWRStaleKey', () => {
+  it('should create control key', () => {
+    expect(RedisController.createSWRControlKey('key')).toEqual(
+      'key:swr.control'
+    )
+  })
+})
+
+describe('getEntry', () => {
+  it('should call cache.get from service to fetch a key', () => {
+    const service = {
+      cache: {
+        get: key => Promise.resolve(`${key}:value`)
+      }
+    }
+    return RedisController.getEntry(service, 'test').then(result => {
+      expect(result).toEqual('test:value')
+    })
+  })
+})
+
+describe('getSWRControlEntry', () => {
+  it('should return true if key exists with valid wrapper ', () => {
+    const service = {
+      cache: {
+        get: jest.fn(key => Promise.resolve('SWR-CONTROL'))
+      }
+    }
+    return RedisController.getSWRControlEntry(service, 'test').then(result => {
+      expect(service.cache.get).toBeCalledWith('test:swr.control')
+      expect(result).toEqual('SWR-CONTROL')
+    })
+  })
+
+  it('should return true if key exists with valid wrapper ', () => {
+    const service = {
+      cache: {
+        get: jest.fn(key => Promise.resolve(undefined))
+      }
+    }
+    return RedisController.getSWRControlEntry(service, 'test').then(result => {
+      expect(service.cache.get).toBeCalledWith('test:swr.control')
+      expect(result).toEqual(undefined)
+    })
+  })
+})
+
+describe('setEntry', () => {
+  it('should pass all arguments to cache.set', () => {
+    const service = {
+      cache: {
+        set: jest.fn()
+      }
+    }
+    RedisController.setEntry(service, 'key', 'value', 200)
+    expect(service.cache.set).toBeCalledWith('key', 'value', 200)
+  })
+})
+
+describe('setSWRStaleEntry', () => {
+  it('should create a key that has no ttl', () => {
+    const service = {
+      cache: {
+        set: jest.fn()
+      }
+    }
+    RedisController.setSWRStaleEntry(service, 'key', 'value', 100)
+    expect(service.cache.set).toBeCalledWith('key:swr.stale', 'value', 100)
+  })
+})
+
+describe('setSWRControlEntry', () => {
+  it('should create a key that has no ttl', () => {
+    const service = {
+      cache: {
+        set: jest.fn()
+      }
+    }
+    RedisController.setSWRControlEntry(service, 'key')
+    expect(service.cache.set).toBeCalledWith(
+      'key:swr.control',
+      undefined,
+      undefined
+    )
+  })
+
+  it('should create a key that with ttl and value', () => {
+    const service = {
+      cache: {
+        set: jest.fn()
+      }
+    }
+    RedisController.setSWRControlEntry(service, 'key', 200, 'SWR-CONTROL')
+    expect(service.cache.set).toBeCalledWith(
+      'key:swr.control',
+      'SWR-CONTROL',
+      200
+    )
+  })
+})
